+++ src/citra_qt/configuration/configure_dialog.h
