+++ src/citra_qt/configuration/configure_general.ui
