+++ src/citra_qt/configuration/configure_graphics.ui
