+++ src/citra_qt/configuration/configure_input.cpp
@@ -6,8 +6,8 @@
 #include <memory>
 #include <utility>
 #include <QTimer>
-#include "citra_qt/config.h"
-#include "citra_qt/configure_input.h"
+#include "citra_qt/configuration/config.h"
+#include "citra_qt/configuration/configure_input.h"
 #include "common/param_package.h"
 #include "input_common/main.h"
